requirements.txt
graia-ariadne
graia-saya
graia-amnesia
graia-scheduler
graiax-shortcut
json5
opencv-python
pillow